+++ Runtime/Scripts/Manipulation/UxrPlacementOptions.cs
@@ -1,24 +1,32 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="UxrPlacementType.cs" company="VRMADA">
+// <copyright file="UxrPlacementOptions.cs" company="VRMADA">
 //   Copyright (c) VRMADA, All rights reserved.
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
+using System;
+
 namespace UltimateXR.Manipulation
 {
     /// <summary>
     ///     Enumerates the different ways a <see cref="UxrGrabbableObject" /> can transition when being placed on an
     ///     <see cref="UxrGrabbableObjectAnchor" />.
     /// </summary>
-    public enum UxrPlacementType
+    [Flags]
+    public enum UxrPlacementOptions
     {
         /// <summary>
-        ///     Place immediately.
+        ///     Place immediately. If the object is being grabbed, release it.
         /// </summary>
-        Immediate,
+        None = 0,
 
         /// <summary>
         ///     Place using smooth transition (interpolation).
         /// </summary>
-        Smooth
+        Smooth = 1,
+
+        /// <summary>
+        ///     Do not release the object when placing.
+        /// </summary>
+        DontRelease = 1 << 1
     }
 }